Rename imageops sources.
