+++ src/imagefileops.hpp
