+++ src/meson.build
@@ -5,7 +5,7 @@ capypdf_lib = shared_library('capypdf',
   'pdfgen.cpp',
   'pdfdrawcontext.cpp',
   'pdfdocument.cpp',
-  'imageops.cpp',
+  'imagefileops.cpp',
   'utils.cpp',
   'pdfcolorconverter.cpp',
   'fontsubsetter.cpp',
